add tensorboard

+++ model/modules/trainer_nn.py
@@ -15,14 +15,30 @@ from utils.dataloader import SeqDataset
 from utils.loss import L1regularization
 from opts import model_opts, DATASET
 from modules.model_ae import AutoEncoder, UnbAutoEncoder, Decoder
+from tensorboardX import SummaryWriter
 
 class TrainProcess():
     def __init__(self, args):
         self.args = args
-        self.device = torch.device('cuda:{}'.format(args.gpu_ids[0])) if args.gpu_ids else torch.device('cpu')
+        self.writer = SummaryWriter(log_dir=f'../../runs/{args.exp_name}')
+        self.device = torch.device(f'cuda:{args.gpu_ids[0]}') if args.gpu_ids else torch.device('cpu')
 
-        self.trainset = SeqDataset(DATASET[args.mode]['train_mod1'], DATASET[args.mode]['train_mod2'])
-        self.testset = SeqDataset(DATASET[args.mode]['test_mod1'])
+        if args.selection:
+            self.trainset = SeqDataset(
+                DATASET[args.mode]['train_mod1'], DATASET[args.mode]['train_mod2'],
+                mod1_idx_path=args.mod1_idx_path, mod2_idx_path=args.mod2_idx_path
+                )
+            self.testset = SeqDataset(
+                DATASET[args.mode]['test_mod1'], DATASET[args.mode]['test_mod2'], 
+                mod1_idx_path=args.mod1_idx_path, mod2_idx_path=args.mod2_idx_path
+                )
+
+            args.mod1_dim = args.select_dim if args.mod1_idx_path != None else args.mod1_dim
+            args.mod2_dim = args.select_dim if args.mod2_idx_path != None else args.mod2_dim
+            
+        else:
+            self.trainset = SeqDataset(DATASET[args.mode]['train_mod1'], DATASET[args.mode]['train_mod2'])
+            self.testset = SeqDataset(DATASET[args.mode]['test_mod1'], DATASET[args.mode]['test_mod2'])
 
         self.train_loader = DataLoader(self.trainset, batch_size=args.batch_size, shuffle=True)
         self.test_loader = DataLoader(self.testset, batch_size=args.batch_size, shuffle=False)
@@ -69,6 +85,7 @@ class TrainProcess():
         self.model.train()
         self.adjust_learning_rate(self.optimizer, epoch)
 
+        total_loss = 0.0
         total_rec_loss = 0.0
         print(f'Epoch {epoch+1:2d} / {self.args.epoch}')
 
@@ -78,25 +95,49 @@ class TrainProcess():
             mod2_seq = mod2_seq.to(self.device).float()
             mod2_rec = self.model(mod1_seq)
             rec_loss = self.mse_loss(mod2_rec, mod2_seq)
-            l1reg_loss = self.l1reg_loss(self.model)
+            l1reg_loss = self.l1reg_loss(self.model) * self.args.reg_loss_weight
 
-            loss = rec_loss + l1reg_loss * self.args.reg_loss_weight
+            loss = rec_loss + l1reg_loss
 
             self.optimizer.zero_grad()
             loss.backward()
             self.optimizer.step()
 
-            total_rec_loss += loss.item()
+            total_loss += loss.item()
+            total_rec_loss += rec_loss.item()
 
+            self.writer.add_scalar("rec_loss", rec_loss.item(), epoch)
             print(f'Epoch {epoch+1:2d} [{batch_idx+1:2d} /{len(self.train_loader):2d}] | ' + \
-                f'Total: {total_rec_loss / (batch_idx + 1):.4f} | ' + \
+                f'Total: {total_loss / (batch_idx + 1):.4f} | ' + \
                 f'Rec: {rec_loss.item():.4f} | ' + \
                 f'L1 Reg: {l1reg_loss.item():.4f}')
+
         
+        train_rmse = np.sqrt(total_rec_loss / len(self.train_loader))
+        print(f'Epoch {epoch+1:3d} / {self.args.epoch} | Train RMSE: {train_rmse:.4f}', end=" ")
+        self.eval_epoch(epoch)
+        self.writer.add_scalar("train_rmse", train_rmse, epoch)
+
         # save checkpoint
-        filename = f"weights/model_{self.args.arch}_{self.args.mode}_{self.args.name}.pt"
-        print(f"saving weight to {filename} ...")
-        torch.save(self.model.state_dict(), filename)
+        if not self.args.dryrun:
+            filename = f"../../weights/model_{self.args.arch}_{self.args.mode}_{self.args.name}.pt"
+            print(f"saving weight to {filename} ...")
+            torch.save(self.model.state_dict(), filename)
+
+    def eval_epoch(self, epoch):
+        self.model.eval()
+
+        total_rec_loss = 0.0
+        for batch_idx, (mod1_seq, mod2_seq) in enumerate(self.test_loader):
+            mod1_seq = mod1_seq.to(self.device).float()
+            mod2_seq = mod2_seq.to(self.device).float()
+            mod2_rec = self.model(mod1_seq)
+            rec_loss = self.mse_loss(mod2_rec, mod2_seq)
+            total_rec_loss += rec_loss.item()
+        test_rmse = np.sqrt(total_rec_loss / len(self.test_loader))
+        print(f'| Eval RMSE: {test_rmse:.4f}')
+        self.writer.add_scalar("test_rmse", test_rmse, epoch)
+
 
     def run(self):
         self.load_checkpoint()
